Show first menu item as icon
